+++ src/game/nlcore/common/elements.ts
@@ -1,7 +1,13 @@
 import {Character} from "../elements/character";
 import {Condition, Lambda} from "../elements/condition";
 import {Control} from "@core/elements/control";
-import {Image as ImageClass, RichImageConfig, TagDefinitions, TagGroupDefinition} from "../elements/image";
+import {
+    Image as ImageClass,
+    RichImageUserConfig,
+    TagDefinitions,
+    TagGroupDefinition,
+    TagSrcResolver
+} from "../elements/displayable/image";
 import {Menu} from "../elements/menu";
 import {Scene} from "../elements/scene";
 import {Script} from "../elements/script";
@@ -10,27 +16,47 @@ import {Story} from "../elements/story";
 import {Transform} from "@core/elements/transform/transform";
 import {Sentence} from "@core/elements/character/sentence";
 import {Word} from "@core/elements/character/word";
-import {Text} from "@core/elements/text";
+import {Text} from "@core/elements/displayable/text";
 import {Pause} from "@core/elements/character/pause";
+import {StaticImageData} from "@core/types";
 
 interface ImageConstructor {
     new<T extends TagGroupDefinition | null>(
-        config: Partial<RichImageConfig<T>> & {
-            tag: T extends TagGroupDefinition ? TagDefinitions<T> : never;
-        }
+        config: Omit<Partial<RichImageUserConfig<T>>, "src"> &
+            (T extends null ?
+                {
+                    src: string | StaticImageData;
+                    tag?: never;
+                } : T extends TagGroupDefinition ?
+                    {
+                        src: TagSrcResolver<T>;
+                        tag: TagDefinitions<T>;
+                    }
+                    : never),
     ): ImageClass<T>;
 }
 
 const Image: ImageConstructor = function <T extends TagGroupDefinition | null>(
     this: ImageClass<T>,
-    config: Partial<RichImageConfig<T>> & {
-        tag: T extends TagGroupDefinition ? TagDefinitions<T> : never;
-    }
+    config: Omit<Partial<RichImageUserConfig<T>>, "src"> &
+        (T extends null ?
+            {
+                src: string | StaticImageData;
+                tag?: never;
+            } : T extends TagGroupDefinition ?
+                {
+                    src: TagSrcResolver<T>;
+                    tag: TagDefinitions<T>;
+                }
+                : never),
 ): ImageClass<T> {
     if (!new.target) {
         throw new Error("Image is a constructor and should be called with new keyword");
     }
-    return new ImageClass<T>(config, config.tag);
+    return new ImageClass<T>(
+        config as Partial<RichImageUserConfig<T>>,
+        config.tag as TagDefinitions<T> | undefined
+    );
 } as unknown as ImageConstructor;
 
 export {