+++ src/game/nlcore/elements/type.ts
@@ -1,4 +1,4 @@
-import {TagGroupDefinition} from "@core/elements/image";
+import {TagGroupDefinition} from "@core/elements/displayable/image";
 
 export type {
     TagGroupDefinition,