+++ src/game/player/elements/preload/Preload.tsx
@@ -4,7 +4,7 @@ import {Sound} from "@core/elements/sound";
 import {SrcManager} from "@core/action/srcManager";
 import {usePreloaded} from "@player/provider/preloaded";
 import {Preloaded, PreloadedSrc} from "@player/lib/Preloaded";
-import {Image as GameImage} from "@core/elements/image";
+import {Image as GameImage} from "@core/elements/displayable/image";
 
 export function Preload(
     {